requirements.txt
python-dotenv 
praw
langchain
langchain_community
streamlit
langchain_groq